+++ crates/gen/src/bool32.rs
