+++ crates/gen/src/bstr.rs
