+++ crates/gen/src/callback.rs
