+++ crates/gen/src/class.rs
