+++ crates/gen/src/delegate.rs
