+++ crates/gen/src/enum.rs
