+++ crates/gen/src/lib.rs
@@ -1,4 +1,3 @@
-mod r#async;
 mod gen;
 mod gen_helpers;
 mod interface_info;
@@ -6,7 +5,26 @@ mod iterator;
 mod object;
 mod propertykey;
 mod to_ident;
-mod types;
+mod bool32;
+mod bstr;
+mod callback;
+mod class;
+mod com_interface;
+mod delegate;
+mod r#enum;
+mod handle;
+mod interface;
+mod matrix3x2;
+mod matrix4x4;
+mod ntstatus;
+mod pstr;
+mod pwstr;
+mod r#struct;
+mod timespan;
+mod vector2;
+mod vector3;
+mod vector4;
+mod r#async;
 
 pub use gen::*;
 pub use gen_helpers::*;
@@ -15,7 +33,25 @@ use iterator::*;
 use object::*;
 use propertykey::*;
 use quote::*;
-use r#async::*;
 use reader::*;
 pub use to_ident::*;
-use types::*;
+pub use bool32::*;
+pub use bstr::*;
+pub use callback::*;
+pub use class::*;
+pub use com_interface::*;
+pub use delegate::*;
+pub use handle::*;
+pub use interface::*;
+pub use matrix3x2::*;
+pub use matrix4x4::*;
+pub use ntstatus::*;
+pub use pstr::*;
+pub use pwstr::*;
+pub use r#enum::*;
+pub use r#struct::*;
+pub use timespan::*;
+pub use vector2::*;
+pub use vector3::*;
+pub use vector4::*;
+pub use r#async::*;
