+++ crates/gen/src/pstr.rs
