+++ crates/gen/src/pwstr.rs
