+++ crates/gen/src/timespan.rs
