--- crates/gen/src/types/mod.rs
@@ -1,41 +1,0 @@
-pub use super::*;
-
-mod bool32;
-mod bstr;
-mod callback;
-mod class;
-mod com_interface;
-mod delegate;
-mod r#enum;
-mod handle;
-mod interface;
-mod matrix3x2;
-mod matrix4x4;
-mod ntstatus;
-mod pstr;
-mod pwstr;
-mod r#struct;
-mod timespan;
-mod vector2;
-mod vector3;
-mod vector4;
-
-pub use bool32::*;
-pub use bstr::*;
-pub use callback::*;
-pub use class::*;
-pub use com_interface::*;
-pub use delegate::*;
-pub use handle::*;
-pub use interface::*;
-pub use matrix3x2::*;
-pub use matrix4x4::*;
-pub use ntstatus::*;
-pub use pstr::*;
-pub use pwstr::*;
-pub use r#enum::*;
-pub use r#struct::*;
-pub use timespan::*;
-pub use vector2::*;
-pub use vector3::*;
-pub use vector4::*;
