+++ crates/gen/src/vector3.rs
